+++ posthog/src/test/java/com/posthog/internal/PostHogFlagsRequestTest.kt
@@ -7,10 +7,10 @@ import com.posthog.groups
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-internal class PostHogDecideRequestTest {
+internal class PostHogFlagsRequestTest {
     @Test
-    fun `sets the decide request content`() {
-        val request = PostHogDecideRequest(API_KEY, DISTINCT_ID, anonymousId = ANON_ID, groups)
+    fun `sets the flags request content`() {
+        val request = PostHogFlagsRequest(API_KEY, DISTINCT_ID, anonymousId = ANON_ID, groups)
 
         assertEquals(API_KEY, request["api_key"])
         assertEquals(DISTINCT_ID, request["distinct_id"])